backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.22.1",
    "apscheduler>=3.11.2",
    "fastapi>=0.136.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "sqlalchemy>=2.0.49",
    "uvicorn[standard]>=0.46.0",
]
